+++ src/Components/routes.js
@@ -2,6 +2,7 @@ import EditScreen from '../Screens/EditScreen';
 import NewManager from '../Screens/NewManager';
 import ListScreen from '../Screens/ListScreen';
 import Doctors from '../Routes/Doctors';
+import Logout from '../Routes/Logout';
 import ProfileManagerScreen from '../Screens/ProfileManagerScreen';
 import {DrawerNavigator} from 'react-navigation';
 
@@ -17,6 +18,9 @@ export default DrawerNavigator({
   },
   'Criar novo gestor': {
     screen: NewManager
+  },
+  Logout: {
+    screen: Logout
   }
 }, {
   drawerWidth: 300
